fix: remove excess files
